+++ src/orionbelt/parser/__init__.py
@@ -1,10 +1,13 @@
 """YAML parsing with line fidelity for OrionBelt Semantic Layer."""
 
 from orionbelt.parser.loader import SourceMap, TrackedLoader
+from orionbelt.parser.merger import ExtendsMerger, MergeError
 from orionbelt.parser.resolver import ReferenceResolver
 from orionbelt.parser.validator import SemanticValidator
 
 __all__ = [
+    "ExtendsMerger",
+    "MergeError",
     "ReferenceResolver",
     "SemanticValidator",
     "SourceMap",